vehicle card functionality, refactoring and optimising routes

+++ src/routes/tripRoutes.js
@@ -3,7 +3,7 @@ const router = express.Router();
 const Trip = require("../models/trip");
 const Vehicle = require("../models/vehicle");
 const { mergeTrips } = require("../services/tripService");
-const { calculateTripDistance } = require("../utility");
+const { formatDate } = require("../utility");
 const WebUser = require("../models/webUser");
 const mongoose = require("mongoose");
 const { loadUser } = require("../middleware/userMiddleware");
@@ -11,20 +11,23 @@ const { loadUser } = require("../middleware/userMiddleware");
 const {
   authenticateAdminApiKey,
   authenticateDriverApiKey,
-  authenticateAnyApiKey,
   authenticateSessionOrApiKey,
 } = require("../services/authenticationService");
 const vehicle = require("../models/vehicle");
 
 // getting all trips
 router.get(
-  "/trips",
+  "/api/trips",
   authenticateSessionOrApiKey,
   loadUser,
   async (req, res) => {
     try {
-      if (req.adminAuthenticated) {
-        const trips = await Trip.find();
+      const { vehicleId: queryVehicleId } = req.query;
+      if (req.isAdminAuthenticated) {
+        // return all trips when using api key
+        const trips = await Trip.find(
+          queryVehicleId ? { vehicleId: queryVehicleId } : {}
+        );
         return res.json(trips);
       }
 
@@ -33,32 +36,33 @@ router.get(
         return res.status(401).json({ message: "Unauthorized Trips access" });
       }
 
-      const sevenDaysAgo = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
-      const matchStage = { startDate: { $gte: sevenDaysAgo } };
+      const thirtyDaysAgo = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
 
-      // Role-specific filters
+      const query = {
+        tripCategory: "business",
+        startTimestamp: { $gte: thirtyDaysAgo },
+        $or: [{ markAsDeleted: false }, { markAsDeleted: { $exists: false } }],
+      };
+
+      // return trips based on user role
+      // but always restrict to last 7 days
       switch (webUser.role) {
         case "dispatcher":
-          matchStage.tripCategory = "business";
+          query.tripCategory = "business";
+          if (queryVehicleId) query.vehicleId = queryVehicleId;
           break;
         case "manager":
-          matchStage.vehicleId = webUser.vehicleId;
+          query.vehicleId = webUser.vehicleId;
           break;
         case "admin":
+          if (queryVehicleId) query.vehicleId = queryVehicleId;
+
           break;
         default:
           return res.status(403).json({ message: "Forbidden" });
       }
 
-      // TODO: maybe change from aggregate to find, but first need to check data-format (string/date)
-      const trips = await Trip.aggregate([
-        {
-          $addFields: {
-            startDate: { $dateFromString: { dateString: "$startTimestamp" } },
-          },
-        },
-        { $match: matchStage },
-      ]);
+      const trips = await Trip.find(query);
 
       res.json(trips);
     } catch (error) {
@@ -66,16 +70,32 @@ router.get(
     }
   }
 );
+
 // getting one trip by id
 router.get("/trip/:id", getTrip, authenticateAdminApiKey, async (req, res) => {
   res.json(res.trip);
 });
 
+// middleware
+async function getTrip(req, res, next) {
+  let trip;
+  try {
+    trip = Trip.findById(req.params.id);
+    if (trip == null) {
+      return res.status(404).json({ message: "Trip not found!" });
+    }
+  } catch (error) {
+    return res.status(500).json({ message: error.message });
+  }
+  res.trip = trip;
+  next();
+}
+
 // creating a trip
-router.post("/trip", authenticateAnyApiKey, async (req, res) => {
-  const timestamp = new Date().toLocaleString();
+router.post("/api/trip", authenticateSessionOrApiKey, async (req, res) => {
+  const timestamp = formatDate(new Date().toLocaleString());
   console.log(req.body);
-  let vehicleId = null;
+  let vehicleId;
   if (req.body.vehicle) {
     const vehicleData = req.body.vehicle;
     try {
@@ -102,18 +122,18 @@ router.post("/trip", authenticateAnyApiKey, async (req, res) => {
   const trip = new Trip({
     startLocation: req.body.startLocation,
     endLocation: req.body.endLocation,
-    startTimestamp: startTimestamp,
-    endTimestamp: endTimestamp,
+    startTimestamp: req.body.startTimestamp,
+    endTimestamp: req.body.endTimestamp,
     startMileage: req.body.startMileage,
     endMileage: req.body.endMileage,
     tripCategory: req.body.tripCategory,
     tripPurpose: req.body.tripPurpose,
     tripNotes: req.body.tripNotes,
     tripStatus: req.body.tripStatus,
-    vehicleId: vehicleData.vin ?? null,
+    vehicleId: vehicleId,
     recorded: req.body.recorded,
   });
-  if (req.body.recorded == false) {
+  if (req.body.recorded === false) {
     trip.tripNotes.push(
       `Die Fahrt konnte aufgrund technischer Probleme nicht aufgezeichnet werden und wurde am ${timestamp} nachgetragen.`
     );
@@ -129,139 +149,49 @@ router.post("/trip", authenticateAnyApiKey, async (req, res) => {
 });
 
 // updating one trip by id
-// TODO: rethink what is editable and what now
-router.patch("/trip/:id", getTrip, async (req, res) => {
+router.patch("/api/trip/:id", getTrip, async (req, res) => {
+  if (!req.authenticatedBySession && !req.authenticatedByApiKey) {
+    return res.status(403).json({ message: "Forbidden" });
+  }
   if (!isTripEditableWithinSevenDays(res.trip)) {
     return res.status(403).json({
       message: "You are not allowed to edit trips older than 7 days!",
     });
   }
 
-  const timestamp = new Date().toLocaleString();
+  const timestamp = formatDate(new Date().toLocaleString());
 
-  if (req.body.checked != null) {
-    res.trip.checked = req.body.checked;
-  }
-  if (req.body.permittedPrivate != null) {
-    res.trip.permittedPrivate = req.body.permittedPrivate;
-  }
-  if (
-    req.body.tripStatus != null &&
-    ["finished", "cancelled"].includes(req.body.tripStatus)
-  ) {
-    if (res.body.tripStatus != null) {
-      res.trip.tripNotes.push(
-        `Der Status wurde am ${timestamp} von *${res.body.tripStatus}* auf *${req.body.tripStatus}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Der Status *${req.body.tripStatus}* hat gefehlt und wurde am ${timestamp} nachgetragen.`
-      );
-    }
-    res.trip.tripStatus = req.body.tripStatus;
-  }
-  if (req.body.startTimestamp != null) {
-    if (res.body.startTimestamp != null) {
-      res.trip.tripNotes.push(
-        `Die Startzeit wurde am ${timestamp} von *${res.body.startTimestamp}* auf *${req.body.startTimestamp}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Die Startzeit *${req.body.startTimestamp}* wurde nicht aufgezeichnet und am ${timestamp} nachgetragen.`
-      );
-    }
-    res.trip.startTimestamp = req.body.startTimestamp;
-  }
-  if (req.body.endTimestamp != null) {
-    if (res.body.endTimestamp != null) {
-      res.trip.tripNotes.push(
-        `Die Endzeit wurde am ${timestamp} von *${res.body.endTimestamp}* auf *${req.body.endTimestamp}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Die Endzeit *${req.body.endTimestamp}* wurde nicht aufgezeichnet und am ${timestamp} nachgetragen.`
-      );
-    }
-    res.trip.endTimestamp = req.body.endTimestamp;
-  }
-  if (req.body.startLocation != null) {
-    if (res.body.startLocation != null) {
-      res.trip.tripNotes.push(
-        `Der Startstandort wurde am ${timestamp} von *${res.body.startLocation}* auf *${req.body.startLocation}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Der Startstandort *${req.body.startLocation}* wurde nicht aufgezeichnet und am ${timestamp} nachgetragen.`
-      );
-    }
-    res.trip.startLocation = req.body.startLocation;
-  }
-  if (req.body.endLocation != null) {
-    if (res.body.endLocation != null) {
-      res.trip.tripNotes.push(
-        `Der Endstandort wurde am ${timestamp} von *${res.body.endLocation}* auf *${req.body.endLocation}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Der Endstandort *${req.body.endLocation}* wurde nicht aufgezeichnet am ${timestamp} nachgetragen.`
-      );
-    }
-    res.trip.endLocation = req.body.endLocation;
+  res.trip.checked = true; // mark as checked
+
+  if (req.body.clientCompany != null) {
+    res.body.clientCompany = req.body.clientCompany;
   }
 
-  if (req.body.startMileage != null) {
-    if (res.body.startMileage != null) {
-      res.trip.tripNotes.push(
-        `Der Startkilometerstand wurde am ${timestamp} von *${res.body.startMileage}* auf *${req.body.startMileage}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Der Startkilometerstand *${req.body.startMileage}* wurde nicht aufgezeichnet und am ${timestamp} nachgetragen.`
-      );
-    }
-    res.trip.startMileage = req.body.startMileage;
+  if (req.body.client != null) {
+    res.body.client = req.body.client;
   }
-  if (req.body.endMileage != null) {
-    if (res.body.endMileage != null) {
-      res.trip.tripNotes.push(
-        `Der EndKilometerstand wurde am ${timestamp} von *${res.body.endMileage}* auf *${req.body.endMileage}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Der Endkilometerstand *${req.body.endMileage}* nicht aufgezeichnet und am ${timestamp} nachgetragen.`
-      );
-    }
-    res.trip.endMileage = req.body.endMileage;
+
+  if (req.body.detourNote != null) {
+    res.trip.detourNote = req.body.detourNote;
   }
+
   if (req.body.tripCategory != null) {
-    if (res.body.tripCategory != null) {
-      res.trip.tripNotes.push(
-        `Die Kategorie wurde am ${timestamp} von *${res.body.tripCategory}* auf *${req.body.tripCategory}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Die Kategorie ${req.body.tripCategory} hat gefehlt und wurde ${timestamp} am hinzugefügt.`
-      );
-    }
+    res.trip.tripNotes.push(
+      `Die Kategorie wurde am ${timestamp} von *${res.body.tripCategory}* auf *${req.body.tripCategory}* korrigiert.`
+    );
     res.trip.tripCategory = req.body.tripCategory;
   }
+
   if (req.body.tripPurpose != null) {
-    if (res.trip.tripPurpose != null) {
-      res.trip.tripNotes.push(
-        `Der Anlass wurde am ${timestamp} von *${res.body.tripPurpose}* auf *${req.body.tripPurpose}* korrigiert.`
-      );
-    } else {
-      res.trip.tripNotes.push(
-        `Der Anlass *${req.body.tripPurpose}* wurde am ${timestamp} hinzugefügt.`
-      );
-    }
-    res.trip.tripPurpose = req.body.tripPurpose;
+    res.tripPurpose = req.body.tripPurpose;
   }
+
   if (req.body.tripNotes != null) {
-    res.trip.tripNotes.push(
-      `Eigene Bemerkung (${timestamp}): ${req.body.tripNotes}`
-    );
+    res.trip.tripNotes.push(`${req.body.tripNotes} | (${timestamp})`);
   }
+
+  // not possible to change data that was recorded
+  // if needed, replace the current trip
 
   try {
     const updatedTrip = await res.trip.save();
@@ -272,7 +202,7 @@ router.patch("/trip/:id", getTrip, async (req, res) => {
 });
 
 // delete multiple trips at once
-router.delete("/trips", authenticateAdminApiKey, async (req, res) => {
+router.delete("/api/trips", authenticateAdminApiKey, async (req, res) => {
   try {
     await Trip.deleteMany();
     res.json({ message: "All trips deleted" });
@@ -282,38 +212,34 @@ router.delete("/trips", authenticateAdminApiKey, async (req, res) => {
 });
 
 // deleting one trip by id
-router.delete("/trip/:id", getTrip, async (req, res) => {
-  try {
-    await res.trip.deleteOne();
-    res.json({ message: "Trip deleted" });
-  } catch (error) {
-    res.status(500).json({ message: error.message });
+router.delete(
+  "/api/trip/:id",
+  authenticateAdminApiKey,
+  getTrip,
+  async (req, res) => {
+    try {
+      if (req.isAdminAuthenticated) {
+        await res.trip.deleteOne();
+        res.json({ message: "Trip deleted from database" });
+      } else {
+        res.trip.markAsDeleted = true;
+        res.json({ message: "Trip marked as deleted" });
+      }
+    } catch (error) {
+      res.status(500).json({ message: error.message });
+    }
   }
-});
+);
 
 // middleware
-async function getTrip(req, res, next) {
-  let trip;
-  try {
-    trip = Trip.findById(req.params.id);
-    if (trip == null) {
-      return res.status(404).json({ message: "Trip not found!" });
-    }
-  } catch (error) {
-    return res.status(500).json({ message: error.message });
-  }
-  res.trip = trip;
-  next();
-}
-// middleware
-// TODO increase length, allow more than two trips
+// TODO test if working
 async function getAllTrips(req, res, next) {
   try {
     const tripIds = req.body.tripIds;
 
-    if (!Array.isArray(tripIds) || tripIds.length !== 2) {
+    if (!Array.isArray(tripIds) || tripIds.length < 2) {
       return res.status(400).json({
-        message: "Exactly two trip IDs must be provided and must be an array.",
+        message: "Please provide at least two trip IDs to merge.",
       });
     }
 
@@ -330,15 +256,14 @@ async function getAllTrips(req, res, next) {
   }
 }
 
-// TODO: allow more than two trips
-router.post("/merge-trips", getAllTrips, async (req, res) => {
+router.post("/api/trips/merge", getAllTrips, async (req, res) => {
   try {
     const trips = res.trips;
-    if (trips.some((trip) => tripStatus === "incorrect")) {
-      return res.status(403).json({
-        message: "You are not allowed to merge invalid trips!",
-      });
-    }
+    // if (trips.some((trip) => tripStatus === "incorrect")) {
+    //   return res.status(403).json({
+    //     message: "You are not allowed to merge invalid trips!",
+    //   });
+    // }
 
     if (trips.some((trip) => !isTripEditableWithinSevenDays(trip))) {
       return res.status(403).json({
@@ -353,7 +278,8 @@ router.post("/merge-trips", getAllTrips, async (req, res) => {
 });
 
 // for export
-router.get("/tripsInRange", async (req, res) => {
+// TODO: check later, add authentication etc.
+router.get("/api/trips/range", async (req, res) => {
   try {
     // user provides ?fromDate=...&toDate=... as query parameters
     const { fromDate, toDate } = req.query;
@@ -368,18 +294,8 @@ router.get("/tripsInRange", async (req, res) => {
 
 function isTripEditableWithinSevenDays(trip) {
   const today = new Date();
-  let tripDate;
-  if (trip.endTimestamp != null) {
-    tripDate = new Date(trip.endTimestamp);
-  } else if (trip.startTimestamp != null) {
-    tripDate = new Date(trip.startTimestamp);
-  } else {
-    tripDate = trip.res.receivedDate;
-  }
+  let tripDate = trip.startTimestamp;
 
-  if (isNaN(tripDate.getTime())) {
-    throw new Error("Invalid trip end timestamp.");
-  }
   const sevenDaysInMs = 7 * 24 * 60 * 60 * 1000;
   const diffInMs = today.getTime() - tripDate.getTime();
   if (diffInMs > sevenDaysInMs) {
